Create README.md
Hello, this is the agreed upon task as per the PDF, my architecture pattern of design is MVP(Model-View-Presenter) in which the UI (view) is dumb and the presenter
acts as the brain of the component handling anything that has logic or networking. I've also managed to add dark mode to the app, couldn't complete the other bonus
requirements since my time was very limited ( 2 days) and I had to manage it between the main requirements. awaiting your feed back and If you have any questions
please contact me at your convenience